API

Base.endswith

Base.startswith

Wrangling.contains

Wrangling.contains

Wrangling.contains_any

Wrangling.contains_any

Wrangling.endswith_any

Wrangling.endswith_any

Wrangling.startswith_any

Wrangling.startswith_any

Wrangling.contains — Method

contains(haystack, needle)

This is the same as Base.occurin(needle, haystack), just with the arguments reversed. This makes it line up with startswith and endswith.

Wrangling.contains — Method

contains(needle) -> Function

Curried form of contains(haystack, needle)

Wrangling.contains_any — Method

contains_any(haystack, needles)

Given a list of needles, this returns true if contains would return true for any needle in that list.

Wrangling.contains_any — Method

contains_any(needles) -> Function

Curried form of contains_any(haystack, needles)

Wrangling.endswith_any — Method

endswith_any(haystack, needles)

Given a list of needles, this returns true if endswith would return true for any needle in that list.

Wrangling.endswith_any — Method

endswith_any(needles) -> Function

Curried form of endswith_any(haystack, needles)

Wrangling.startswith_any — Method

startswith_any(haystack, needles)

Given a list of needles, this returns true if startswith would return true for any needle in that list.

Wrangling.startswith_any — Method

startswith_any(needles) -> Function

Curried form of startswith_any(haystack, needles)

Base.endswith — Method

endswith(needle) -> Function

Curried form of endswith(haystack, needle)

Base.startswith — Method

startswith(needle) -> Function

Curried form of startswith(haystack, needle)